requirements.txt
# installed mosquitto as broker
paho-mqtt # paho-mqtt as client
protobuf # protobuf as message format